MNT: Add configuration option to determine which camera interface to use
