+++ src/detector/implementations/__init__.py
@@ -1,2 +1,7 @@
+from .abstract_camera_interface import AbstractCameraInterface
 from .abstract_marker_interface import AbstractMarkerInterface
 from .aruco_marker_implementation import ArucoMarker
+# Windows may not be able to load picamera2,
+# so don't automatically import all implementations here.
+# Instead, user should directly import the concrete
+# AbstractCameraInterface implementation from the module.
